JPA에서 제공하는 count를 사용하는 것으로 변환
쿼리를 직접 사용하여 날리는것과 다른 부분은 없지만 제공하는 기능을 사용하는 방법으로 변환하였다

+++ src/main/java/com/hosu/sns/controller/PostController.java
@@ -56,7 +56,7 @@ public class PostController {
     }
 
     @GetMapping("/{postId}/likes")
-    public Response<Integer> getLikes(@PathVariable Integer postId, Authentication authentication){
+    public Response<Long> getLikes(@PathVariable Integer postId, Authentication authentication){
         return Response.success(postService.likeCount(postId));
     }
